refactor: minor updates
