+++ src/custom-shapes/graph/graph.ts
@@ -1,10 +1,9 @@
 import type { Editor, TLArrowBinding, TLShapeId, TLUnknownShape } from "tldraw";
 import { DepGraph } from "dependency-graph";
-import { isComponentShape, type Node } from "./types";
-import type { IComponentShape } from "../component-shape/component-shape";
+import { isComponentShape, type GraphNode } from "./types";
 
 export const buildGraph = (startingShapeId: TLShapeId, editor: Editor) => {
-  const graph = new DepGraph<Node>();
+  const graph = new DepGraph<GraphNode>();
   const visited = new Set<string>();
 
   function walk(shapeId: TLShapeId) {
@@ -73,77 +72,4 @@ export const buildGraph = (startingShapeId: TLShapeId, editor: Editor) => {
 
   walk(startingShapeId);
   return graph;
-};
-
-export const getInputsOfNode = (id: string, graph: DepGraph<Node>) => {
-  return graph.dependenciesOf(id).filter((n) => {
-    const node = graph.getNodeData(n);
-    return node;
-  });
-};
-
-export const getOutputsOfNode = (id: string, graph: DepGraph<Node>) => {
-  return graph.dependantsOf(id).filter((n) => {
-    const node = graph.getNodeData(n);
-    return node;
-  });
-};
-
-export const runGraphFromShape = (shapeId: TLShapeId, editor: Editor) => {
-  const graph = buildGraph(shapeId, editor);
-  const processedNodes = new Set<string>();
-  const queue: string[] = [shapeId];
-
-  while (queue.length > 0) {
-    const currentId = queue.shift()!;
-    if (processedNodes.has(currentId)) continue;
-    const currentNode = graph.getNodeData(currentId);
-    if (!currentNode) continue;
-
-    const inputs = getInputsOfNode(currentId, graph).map((n) =>
-      graph.getNodeData(n)
-    );
-    const outputs = graph
-      .directDependantsOf(currentId)
-      .map((n) => graph.getNodeData(n));
-
-    switch (currentNode.component) {
-      case "instruction":
-        console.log("Running instruction", {
-          instruction: currentNode.value,
-          inputs: inputs
-            .filter((n) => n.component === "text")
-            .map((n) => n.value)
-            .join(", "),
-          outputTypes: [
-            ...new Set(
-              graph
-                .directDependantsOf(currentId)
-                .map<IComponentShape | undefined>((n) =>
-                  editor.getShape(n as TLShapeId)
-                )
-                .filter<IComponentShape>((n) => !!n)
-                .map((n) => n.props.component)
-            ),
-          ],
-          outputNodes: outputs,
-        });
-        processedNodes.add(currentId);
-        break;
-      case "text":
-        console.log("Running text:", currentNode);
-        processedNodes.add(currentId);
-        break;
-      default:
-        processedNodes.add(currentId);
-        break;
-    }
-
-    // Enqueue unprocessed instruction nodes that are direct outputs
-    for (const outputNode of outputs) {
-      if (outputNode && !queue.includes(outputNode.shapeId)) {
-        queue.push(outputNode.shapeId);
-      }
-    }
-  }
 };
